+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 2.0.1-SNAPSHOT
+- Fix chunk light not properly unloading in tall worlds (#28)
 
 ### 2.0.0
 - Update to Minecraft 1.17